feat: open a prefilled topic modal for the slack command
