+++ backend/src/slack/setup.ts
@@ -1,12 +1,15 @@
 import { Express } from "express";
 
+import { setupViews } from "~backend/src/slack/views";
+
 import { slackApp, slackReceiver } from "./app";
-import { setupSlackCommands } from "./commands";
-import { setupSlackShortcuts } from "./shortcuts";
+import { setupCommands } from "./commands";
+import { setupShortcuts } from "./shortcuts";
 
 export function setupSlack(app: Express) {
   app.use("/api", slackReceiver.router);
 
-  setupSlackShortcuts(slackApp);
-  setupSlackCommands(slackApp);
+  setupViews(slackApp);
+  setupShortcuts(slackApp);
+  setupCommands(slackApp);
 }
